+++ src/main/java/org/example/clinicservice/service/impl/MedicalRecordServiceImpl.java
@@ -2,6 +2,12 @@ package org.example.clinicservice.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.example.clinicservice.entity.MedicalRecord;
+import org.example.clinicservice.entity.PatientVisitHistory;
+import org.example.clinicservice.exceptions.ErrorMessage;
+import org.example.clinicservice.exceptions.medicalRecordExceptions.MedicalRecordAlreadyExistsException;
+import org.example.clinicservice.exceptions.medicalRecordExceptions.MedicalRecordNotFoundException;
+import org.example.clinicservice.exceptions.medicalRecordExceptions.MedicalRecordSaveException;
+import org.example.clinicservice.exceptions.medicalRecordExceptions.MedicalRecordsNotFoundException;
 import org.example.clinicservice.repository.MedicalRecordRepository;
 import org.example.clinicservice.repository.SpecialistRepository;
 import org.example.clinicservice.service.interfeces.MedicalRecordService;
@@ -19,44 +25,56 @@ public class MedicalRecordServiceImpl implements MedicalRecordService {
     @Override
     public MedicalRecord getMedicalRecordById(UUID recordId) {
 
-        return medicalRecordRepository.findByRecordId(recordId);
+        try {
+            MedicalRecord medicalRecord = medicalRecordRepository.findByRecordId(recordId);
+            if(recordId == null){
+                throw new MedicalRecordNotFoundException(ErrorMessage.MEDICAL_RECORD_NOT_FOUND);
+            }
+            return medicalRecord;
+        } catch (Exception e) {
+            throw new RuntimeException("An error occurred while retrieving medical record with id " + recordId, e);
+        }
     }
 
     @Override
     public List<MedicalRecord> getMedicalRecordsByPatientId(UUID patientId) {
 
-        return medicalRecordRepository.findByPatientId(patientId);
-    }
-
-    @Override
-    public List<MedicalRecord> getMedicalRecordsByDoctorId(UUID doctorId) {
-
-        return medicalRecordRepository.findByDoctorId(doctorId);
-    }
-
-    @Override
-    public List<MedicalRecord> searchMedicalRecordsByPrescription(String prescription) {
-
-        return medicalRecordRepository.findByPrescriptionContaining(prescription);
-    }
-
-    @Override
-    public List<MedicalRecord> findByDiagnoseContaining(String diagnose) {
-        return medicalRecordRepository.findByDiagnoseContaining(diagnose);
-    }
-
-    @Override
-    public List<MedicalRecord> findByDoctorConclusionContaining(String doctorConclusion) {
-        return medicalRecordRepository.findByDoctorConclusionContaining(doctorConclusion);
+        try {
+            List<MedicalRecord> medicalRecords = medicalRecordRepository.findByPatientId(patientId);
+            if (medicalRecords.isEmpty()) {
+                throw new MedicalRecordsNotFoundException(ErrorMessage.MEDICAL_RECORDS_NOT_FOUND);
+            }
+            return medicalRecords;
+        } catch (Exception e) {
+            throw new RuntimeException("Failed to retrieve medical records for patient ID: " + patientId, e);
+        }
     }
 
     @Override
     public void saveMedicalRecord(MedicalRecord medicalRecord) {
-        medicalRecordRepository.save(medicalRecord);
+
+        if (medicalRecordRepository.existsById(medicalRecord.getRecordId())){
+            throw new MedicalRecordAlreadyExistsException(ErrorMessage.MEDICAL_RECORD_ALREADY_EXISTS);
+        }
+        try {
+            medicalRecordRepository.save(medicalRecord);
+        } catch (Exception e) {
+            throw new MedicalRecordSaveException("Failed to save medical record due to an unexpected error: " + e.getMessage());
+        }
     }
 
     @Override
     public void deleteMedicalRecord(UUID recordId) {
-        medicalRecordRepository.deleteById(recordId);
+
+        if (!medicalRecordRepository.existsById(recordId)) {
+            throw new MedicalRecordNotFoundException(ErrorMessage.MEDICAL_RECORD_NOT_FOUND);
+        }
+
+        try {
+            medicalRecordRepository.deleteById(recordId);
+        } catch (IllegalArgumentException e) {
+            throw new RuntimeException("Failed to delete medical record with ID: " + recordId, e);
+        }
     }
-}
+    }
+
