+++ src/main/java/org/example/clinicservice/service/impl/ScheduleServiceImpl.java
@@ -2,15 +2,20 @@ package org.example.clinicservice.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.example.clinicservice.entity.Schedule;
+import org.example.clinicservice.entity.Specialist;
+import org.example.clinicservice.exceptions.ErrorMessage;
+import org.example.clinicservice.exceptions.scheduleExceptions.ScheduleAlreadyExistsException;
+import org.example.clinicservice.exceptions.scheduleExceptions.ScheduleNotFoundException;
 import org.example.clinicservice.repository.ScheduleRepository;
 import org.example.clinicservice.repository.SpecialistRepository;
 import org.example.clinicservice.service.interfeces.ScheduleService;
 import org.springframework.stereotype.Service;
 
 import java.time.DayOfWeek;
-import java.time.LocalTime;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -21,45 +26,81 @@ public class ScheduleServiceImpl implements ScheduleService {
     @Override
     public Schedule getScheduleById(UUID scheduleId) {
 
-        return scheduleRepository.findByScheduleId(scheduleId);
+        try {
+            Schedule schedule = scheduleRepository.findByScheduleId(scheduleId);
+            if(schedule == null){
+                throw new ScheduleNotFoundException(ErrorMessage.SCHEDULE_NOT_FOUND);
+            }
+            return schedule;
+        } catch (Exception e) {
+            throw new RuntimeException("An error occurred while retrieving schedule with id " + scheduleId, e);
+        }
     }
 
     @Override
-    public List<Schedule> getSchedulesByDoctorId(UUID doctorId) {
+    public List<Schedule> getSchedulesByDoctor(UUID specialistId) {
 
-        return scheduleRepository.findByDoctorId(doctorId);
+        List<Schedule> schedules = scheduleRepository.findBySpecialistId(specialistId);
+
+        if (schedules.isEmpty()) {
+            throw new ScheduleNotFoundException("No schedules found for specialist with ID " + specialistId);
+        }
+
+        return schedules;
     }
 
-    @Override
-    public List<Schedule> getSchedulesByWorkingDay(DayOfWeek dayOfWeek) {
-
-        return scheduleRepository.findByWorkingDaysContaining(dayOfWeek);
-    }
 
     @Override
-    public List<Schedule> getSchedulesByStartTimeBetween(LocalTime startTime, LocalTime endTime) {
+    public Map<UUID, List<DayOfWeek>> getSchedulesByWorkingDay(DayOfWeek dayOfWeek) {
 
-        return scheduleRepository.findByStartTimeBetweenAndEndTime(startTime, endTime);
+        List<Schedule> schedules = scheduleRepository.findByWorkingDaysContaining(dayOfWeek);
+
+        if (schedules.isEmpty()) {
+            throw new ScheduleNotFoundException(ErrorMessage.SCHEDULE_NOT_FOUND_FOR_DAY + dayOfWeek);
+        }
+
+        return schedules.stream()
+                .collect(Collectors.toMap(
+                        schedule -> schedule.getDoctor().getSpecialistId(),
+                        Schedule::getWorkingDays
+                ));
     }
 
     @Override
     public List<Schedule> getSchedulesByDoctorIdAndWorkingDay(UUID doctorId, DayOfWeek dayOfWeek) {
 
-        return scheduleRepository.findByDoctorIdAndWorkingDaysContaining(doctorId, dayOfWeek);
-    }
+        List<Schedule> schedules = scheduleRepository.findByDoctorIdAndWorkingDaysContaining(doctorId, dayOfWeek);
 
-    @Override
-    public List<Schedule> getSchedulesByDoctorIdAndStartTimeBetween(UUID doctorId, LocalTime startTime, LocalTime endTime) {
-        return scheduleRepository.findByDoctorIdAndStartTimeBetween(doctorId, startTime, endTime);
+        if (schedules.isEmpty()) {
+            throw new ScheduleNotFoundException(ErrorMessage.SCHEDULE_NOT_FOUND_FOR_DOCTOR_AND_DAY);
+        }
+        return schedules;
     }
 
     @Override
     public void saveSchedule(Schedule schedule) {
-        scheduleRepository.save(schedule);
+
+        if (scheduleRepository.existsByDoctorAndWorkingDays(schedule.getDoctor(), schedule.getWorkingDays())) {
+            throw new ScheduleAlreadyExistsException(ErrorMessage.SCHEDULE_ALREADY_EXISTS);
+        }
+        try {
+            scheduleRepository.save(schedule);
+        } catch (IllegalArgumentException e) {
+            throw new RuntimeException("Failed to save schedule due to invalid argument: " + e.getMessage(), e);
+        }
     }
 
     @Override
     public void deleteSchedule(UUID scheduleId) {
-        scheduleRepository.deleteById(scheduleId);
+
+        if (!scheduleRepository.existsById(scheduleId)) {
+            throw new ScheduleNotFoundException(ErrorMessage.SCHEDULE_NOT_FOUND);
+        }
+
+        try {
+            scheduleRepository.deleteById(scheduleId);
+        } catch (IllegalArgumentException e) {
+            throw new RuntimeException("Failed to delete schedule with ID: " + scheduleId, e);
+        }
     }
 }
